console/cmd: follow libtrx folder structure
