+++ meson.build
@@ -101,8 +101,26 @@ sources = [
   'src/game/carrier.c',
   'src/game/clock.c',
   'src/game/collide.c',
-  'src/game/console.c',
-  'src/game/console_cmd.c',
+  'src/game/console/cmd/braid.c',
+  'src/game/console/cmd/cheats.c',
+  'src/game/console/cmd/die.c',
+  'src/game/console/cmd/end_level.c',
+  'src/game/console/cmd/exit_game.c',
+  'src/game/console/cmd/exit_to_title.c',
+  'src/game/console/cmd/flipmap.c',
+  'src/game/console/cmd/fly.c',
+  'src/game/console/cmd/fps.c',
+  'src/game/console/cmd/kill.c',
+  'src/game/console/cmd/load_game.c',
+  'src/game/console/cmd/play_demo.c',
+  'src/game/console/cmd/play_level.c',
+  'src/game/console/cmd/save_game.c',
+  'src/game/console/cmd/speed.c',
+  'src/game/console/cmd/teleport.c',
+  'src/game/console/cmd/vsync.c',
+  'src/game/console/cmd/wireframe.c',
+  'src/game/console/common.c',
+  'src/game/console/setup.c',
   'src/game/creature.c',
   'src/game/effect_routines/bubbles.c',
   'src/game/effect_routines/chain_block.c',
