+++ src/game/console/common.h
