+++ src/game/phase/phase_inventory.c
@@ -2,7 +2,7 @@
 
 #include "config.h"
 #include "game/clock.h"
-#include "game/console.h"
+#include "game/console/common.h"
 #include "game/game.h"
 #include "game/gameflow.h"
 #include "game/input.h"
